testing crud reservation and ticket and implementing mail

+++ backend/routers/reservationRouter.js
@@ -3,25 +3,8 @@ const router = express.Router();
 const Reservation = require('../models/ReservationModel');
 const Ticket = require('../models/TicketModel');
 const User = require('../models/UserModel');
-const nodemailer = require('nodemailer');
 const { USER } = require('../constants/userEnum');
-
-const prepareMail = async () => {
-  return await nodemailer.createTestAccount();
-};
-
-let testAccount = prepareMail();
-
-let transporter = nodemailer.createTransport({
-  host: 'smtp.ethereal.email',
-  port: 587,
-  secure: false,
-  auth: {
-    user: testAccount.user,
-    pass: testAccount.pass,
-  },
-});
-
+const sendMail = require('../controllers/mailSender');
 //TODO : reimplement sanatizeData
 const sanatizeData = (data) => {
   ['creatorId', '_id'].forEach((f) => delete data[f]);
@@ -30,13 +13,14 @@ const sanatizeData = (data) => {
 
 const joinReservationAndTicket = async (reservations) => {
   const result = [];
-  for (let reservation in reservations) {
+  for (let reservation of reservations) {
     let tickets = [];
     for (let _id of reservation.tickets) {
       let ticket = await Ticket.findOne({ _id });
       tickets.push(ticket);
     }
     result.push({
+      _id: reservation._id,
       userId: reservation.userId,
       tickets,
     });
@@ -44,30 +28,19 @@ const joinReservationAndTicket = async (reservations) => {
   return result;
 };
 
-router.use((req, res, next) => {
-  if (req.userData) {
-    next();
-    return;
-  } else {
-    res.status(401).json({
-      success: false,
-      msg: 'unautherized access',
-    });
-  }
-});
-
 router.get('/reservations/:page', async (req, res) => {
   try {
     let page = Number(req.params.page);
     let reservation = null;
-    if (req.userData.role === USER)
+    if (req.userData.role === USER) {
       reservation = await Reservation.find({ userId: req.userData.id })
         .skip((page - 1) * 20)
         .limit(20);
-    else
+    } else {
       reservation = await Reservation.find()
         .skip((page - 1) * 20)
         .limit(20);
+    }
     reservation = await joinReservationAndTicket(reservation);
 
     res.status(200).json({
@@ -86,7 +59,7 @@ router.get('/reservations/:page', async (req, res) => {
 
 router.get('/reservation/:reservationId', async (req, res) => {
   try {
-    let _id = Number(req.params.page);
+    let _id = req.params.reservationId;
     let reservation = null;
     if (req.userData.role === USER)
       reservation = await Reservation.find({ _id, userId: req.userData.id });
@@ -110,9 +83,9 @@ router.get('/reservation/:reservationId', async (req, res) => {
 router.post('/reservation', async (req, res) => {
   try {
     const tickets = [];
-    for (let reservation in req.body.reservations) {
-      let ticket = reservation.flight[0];
-      for (let seat in reservation.seats) {
+    for (let reservation of req.body.reservations) {
+      let ticket = reservation.flight;
+      for (let seat of reservation.seats) {
         ticket.seatNumber = seat;
         let result = await Ticket.create(ticket);
         tickets.push(result._id);
@@ -162,26 +135,27 @@ router.put('/reservation/:reservationId', async (req, res) => {
 router.delete('/reservation/:reservationId', async (req, res) => {
   try {
     const _id = req.params.reservationId;
-    let result = [];
-    let mail = User.findOne();
+    let result = {};
+    let { email } = await User.findOne({ _id: req.userData.id });
 
-    if (req.userData.role === USER)
+    if (req.userData.role === USER) {
       result = await Reservation.findOneAndDelete({
         _id,
         userId: req.userData.id,
       });
-    else result = await Reservation.findOneAndDelete({ _id });
-
-    for (let ticket in result.tickets) {
-      await Ticket.findOneAndDelete({ _id: ticket });
+    } else {
+      result = await Reservation.findOneAndDelete({ _id });
     }
+    if (result)
+      for (let ticket of result.tickets) {
+        await Ticket.findOneAndDelete({ _id: ticket });
+      }
 
-    await transporter.sendMail({
-      from: '"Chad Airlines" <airlineschad@gmail.com>',
-      to: mail,
-      subject: 'Cancel reservation',
-      text: `You canceled your reservation to the tickets`,
-    });
+    await sendMail(
+      email,
+      'Cancel reservation',
+      'you have cancelled your reservation'
+    );
 
     res.status(200).json({
       success: true,
