+++ backend/routers/ticketRouter.js
@@ -2,25 +2,9 @@ const express = require('express');
 const router = express.Router();
 const Ticket = require('../models/TicketModel');
 const Reservation = require('../models/ReservationModel');
-const nodemailer = require('nodemailer');
+const sendMail = require('../controllers/mailSender');
 const User = require('../models/UserModel');
-const { ADMIN, USER } = require('../constants/userEnum');
-
-const prepareMail = async () => {
-  return await nodemailer.createTestAccount();
-};
-
-let testAccount = prepareMail();
-
-let transporter = nodemailer.createTransport({
-  host: 'smtp.ethereal.email',
-  port: 587,
-  secure: false,
-  auth: {
-    user: testAccount.user,
-    pass: testAccount.pass,
-  },
-});
+const { ADMIN } = require('../constants/userEnum');
 
 //TODO : reimplement sanatizeData
 const sanatizeData = (data) => {
@@ -28,31 +12,20 @@ const sanatizeData = (data) => {
   return data;
 };
 
-router.use((req, res, next) => {
-  if (
-    req.userData?.role === ADMIN ||
-    (req.userData?.role === USER &&
-      !req.url.includes('tickets') &&
-      !req.method === 'POST')
-  ) {
-    next();
-    return;
-  } else {
-    res.status(401).json({
-      success: false,
-      msg: 'unautherized access',
-    });
-  }
-});
-
 //get req to view all the tickets
 router.get('/tickets/:page', async (req, res) => {
   try {
     const page = Number(req.params.page);
-    let tickets = await Ticket.find()
-      .skip((page - 1) * 20)
-      .limit(20);
-
+    let tickets = [];
+    if (req.userData.role === ADMIN)
+      tickets = await Ticket.find()
+        .skip((page - 1) * 20)
+        .limit(20);
+    else {
+      tickets = await Ticket.find({ userId: req.userData.id })
+        .skip((page - 1) * 20)
+        .limit(20);
+    }
     res.status(200).json({
       success: true,
       msg: 'ok',
@@ -116,41 +89,42 @@ router.delete('/ticket/:ticketId', async (req, res) => {
   try {
     let permission = false;
     let deletedTickets = [];
-    let mail = await User.find({ _id: req.userData.id });
+    let { email } = await User.findOne({ _id: req.userData.id });
 
     const _id = req.params.ticketId;
-    const reservation = await Reservation.find({ tickets: [_id] });
-
-    for (let ticketId in reservation.tickets) {
-      if (ticketId != _id) {
-        let ticket = await Ticket.findOne({ _id: ticketId });
-        if (!ticket.isChild) {
-          permission = true;
+    let reservation = null;
+    if (req.userData.role === ADMIN)
+      reservation = await Reservation.findOne({ tickets: _id });
+    else
+      reservation = await Reservation.findOne({
+        userId: req.userData.id,
+        tickets: _id,
+      });
+    if (reservation) {
+      for (let ticketId of reservation.tickets) {
+        if (ticketId != _id) {
+          let ticket = await Ticket.findOne({ _id: ticketId });
+          if (!ticket.isChild) {
+            permission = true;
+          }
         }
       }
-    }
 
-    if (permission) {
-      deletedTickets.push(await Ticket.findOneAndDelete({ _id }));
-      await transporter.sendMail({
-        from: '"Chad Airlines" <airlineschad@gmail.com>',
-        to: mail,
-        subject: 'Cancel ticket',
-        text: `You canceled your ticket`,
-      });
-    } else {
-      const result = await Reservation.findOneAndDelete(reservation);
+      if (permission) {
+        deletedTickets.push(await Ticket.findOneAndDelete({ _id }));
+        await sendMail(email, 'Cancel ticket', `You canceled your ticket`);
+      } else {
+        const result = await Reservation.findOneAndDelete(reservation);
+        for (let ticket of result.tickets) {
+          deletedTickets.push(await Ticket.findOneAndDelete({ _id: ticket }));
+        }
 
-      for (let ticket in result.tickets) {
-        deletedTickets.push(await Ticket.findOneAndDelete({ _id: ticket }));
+        await sendMail(
+          email,
+          'Cancel reservation',
+          `You canceled your reservation to the tickets`
+        );
       }
-
-      await transporter.sendMail({
-        from: '"Chad Airlines" <airlineschad@gmail.com>',
-        to: mail,
-        subject: 'Cancel reservation',
-        text: `You canceled your reservation to the tickets`,
-      });
     }
     res.status(200).json({
       success: true,
